+++ indispensable/src/commonTest/kotlin/SymmetricEncryptionTest.kt
@@ -6,14 +6,20 @@ import at.asitplus.signum.indispensable.symmetric.SymmetricKey
 import at.asitplus.signum.indispensable.symmetric.preferredMacKeyLength
 import at.asitplus.signum.indispensable.symmetric.randomKey
 
-import io.kotest.core.spec.style.FreeSpec
+import at.asitplus.testballoon.invoke
+import at.asitplus.testballoon.withData
+import de.infix.testBalloon.framework.testSuite
 import io.kotest.datatest.withData
 import io.kotest.matchers.shouldBe
+import kotlin.random.Random
+import de.infix.testBalloon.framework.TestConfig
+import kotlin.time.Duration.Companion.minutes
+import de.infix.testBalloon.framework.testScope
 
-class SymmetricEncryptionTest: FreeSpec({
+val SymmetricEncryptionTest by testSuite(testConfig = TestConfig.testScope(isEnabled = true, timeout = 20.minutes)) {
    
     withData(nameFn={ "Key generation: $it" }, SymmetricEncryptionAlgorithm.entries) { alg ->
-        val key = alg.randomKey()
+        val key = alg.randomKey(randomnessSourceOverride = Random.Default)
 
         key.algorithm shouldBe alg
 
@@ -31,4 +37,4 @@ class SymmetricEncryptionTest: FreeSpec({
             else -> error("unreachable")
         }
     }
-})
+}