+++ supreme/src/androidDeviceTest/kotlin/at/asitplus/signum/supreme/os/AndroidKeyStoreProviderTests.kt
@@ -5,9 +5,11 @@ import at.asitplus.signum.indispensable.SignatureAlgorithm
 import at.asitplus.signum.supreme.sign.verifierFor
 import at.asitplus.signum.supreme.sign.verify
 import at.asitplus.signum.supreme.signature
-import at.asitplus.signum.supreme.succeed
-import br.com.colman.kotest.FreeSpec
-import io.kotest.matchers.should
+import at.asitplus.testballoon.invoke
+import de.infix.testBalloon.framework.TestConfig
+import kotlin.time.Duration.Companion.minutes
+import de.infix.testBalloon.framework.testSuite
+import io.kotest.matchers.shouldBe
 import io.kotest.matchers.types.shouldBeInstanceOf
 import io.kotest.property.Arb
 import io.kotest.property.RandomSource
@@ -15,8 +17,9 @@ import io.kotest.property.arbitrary.Codepoint
 import io.kotest.property.arbitrary.az
 import io.kotest.property.arbitrary.string
 import kotlin.random.Random
+import de.infix.testBalloon.framework.testScope
 
-class AndroidKeyStoreProviderTests : FreeSpec({
+val AndroidKeyStoreProviderTests  by testSuite(testConfig = TestConfig.testScope(isEnabled = true, timeout = 20.minutes)) {
     "Create attested keypair" {
         val alias = Arb.string(minSize = 32, maxSize = 32, Codepoint.az())
             .sample(RandomSource.default()).value
@@ -35,7 +38,7 @@ class AndroidKeyStoreProviderTests : FreeSpec({
         val signature = hardwareSigner.sign(plaintext).signature
 
         SignatureAlgorithm.ECDSAwithSHA256.verifierFor(publicKey).transform {
-            it.verify(plaintext, signature) } should succeed
+            it.verify(plaintext, signature) }.isSuccess shouldBe true
 
     }
-})
+}