+++ supreme/src/androidMain/kotlin/at/asitplus/signum/supreme/sign/EphemeralKeysImpl.kt
@@ -15,6 +15,7 @@ import at.asitplus.signum.indispensable.toCryptoPrivateKey
 import at.asitplus.signum.indispensable.toCryptoPublicKey
 import at.asitplus.signum.indispensable.toJcaPublicKey
 import at.asitplus.signum.indispensable.SecretExposure
+import at.asitplus.signum.indispensable.getJCASignatureInstance
 import at.asitplus.signum.supreme.signCatching
 import com.ionspin.kotlin.bignum.integer.base63.toJavaBigInteger
 import java.security.KeyPair
@@ -31,8 +32,12 @@ actual class EphemeralSignerConfiguration internal actual constructor(): Ephemer
 sealed class AndroidEphemeralSigner (internal val privateKey: PrivateKey) : Signer {
     override val mayRequireUserUnlock = false
     override suspend fun sign(data: SignatureInput) = signCatching {
-        val inputData = data.convertTo(signatureAlgorithm.preHashedSignatureFormat).getOrThrow()
-        signatureAlgorithm.getJCASignatureInstancePreHashed(provider = null).getOrThrow().run {
+        val  (inputData, jcaSignatureInstance ) = when (signatureAlgorithm) {
+            is SignatureAlgorithm.ECDSA ->  data.convertTo(signatureAlgorithm.preHashedSignatureFormat).getOrThrow() to signatureAlgorithm.getJCASignatureInstancePreHashed(provider = null)
+            is SignatureAlgorithm.RSA -> data to signatureAlgorithm.getJCASignatureInstance(provider = null)
+        }
+
+        jcaSignatureInstance.getOrThrow().run {
             initSign(privateKey)
             inputData.data.forEach { update(it) }
             sign().let(::parseFromJca)